PR: move initValues to constants file
